refactor(ai-settings): move AI model block out of Guardrails to Mingo tab
Guardrails no longer shows the LLM Provider / Provider Model card or the
"AI Guardrails" heading. The provider/model selection is extracted into a
self-contained AiModelConfig component (own fetch/save via
useAiConfiguration, save() exposed through a ref) and rendered in the
feature-flagged Mingo AI Chat tab. Guardrails now shows the selected
preset as a read-only summary (title + "Guardrails Preset").
